+++ src/main/java/pl/plajer/buildbattle/commands/MainCommand.java
@@ -1,6 +1,6 @@
 /*
  * BuildBattle - Ultimate building competition minigame
- * Copyright (C) 2019  Plajer's Lair - maintained by Plajer and Tigerpanzer
+ * Copyright (C) 2018  Plajer's Lair - maintained by Plajer and Tigerpanzer
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -61,7 +61,7 @@ public class MainCommand implements CommandExecutor {
 
   public MainCommand(Main plugin) {
     this.plugin = plugin;
-    TabCompletion completion = new TabCompletion();
+    TabCompletion completion = new TabCompletion(plugin);
     plugin.getCommand("buildbattle").setExecutor(this);
     plugin.getCommand("buildbattle").setTabCompleter(completion);
     plugin.getCommand("buildbattleadmin").setExecutor(this);