+++ src/main/java/pl/plajer/buildbattle/menus/options/registry/particles/ParticleRefreshScheduler.java
@@ -1,6 +1,6 @@
 /*
  * BuildBattle - Ultimate building competition minigame
- * Copyright (C) 2019  Plajer's Lair - maintained by Plajer and Tigerpanzer
+ * Copyright (C) 2018  Plajer's Lair - maintained by Plajer and Tigerpanzer
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -41,7 +41,7 @@ public class ParticleRefreshScheduler {
               continue;
             }
             location.getWorld().spawnParticle(buildPlot.getParticles().get(location), location,
-                plugin.getConfig().getInt("Amount-One-Particle-Effect-Contains", 20), 1, 1, 1);
+                plugin.getConfig().getInt("Amount-One-Particle-Effect-Contains", 20), 1, 1, 1, 1);
           }
         }
       }